Point 2
Work_1

+++ src/task1.cpp
@@ -6,12 +6,13 @@ void drawFractal(char(*arr)[M], int x, int y, int size)
 {
 	int space = 0;
 
-	if (size == 0) {
+	if (size == 0) 
+	{
 		arr[x][y] = '*';
 	}
 	else
 	{
-		space = pow(3, size - 1);
+		space = powf(3, size - 1);
 		drawFractal(arr, x, y, size - 1);
 		drawFractal(arr, x + space, y, size - 1);
 		drawFractal(arr, x - space, y, size - 1);
